+++ assets/r/transform_rescale_rank.R
@@ -1,0 +1,14 @@
+# The 'reshape' package provides the 'rescaler' function.
+
+
+# Rescale SELECTED_VAR.
+
+ds[["RRK_SELECTED_VAR"]] <- ds[["SELECTED_VAR"]]
+
+# Convert values to ranks.
+
+ds[["RRK_SELECTED_VAR"]] <-  reshape::rescaler(ds[["SELECTED_VAR"]], "rank")
+
+glimpse(ds)
+summary(ds)
+