+++ lib/features/summary/display.dart
@@ -5,7 +5,7 @@
 /// License: GNU General Public License, Version 3 (the "License")
 /// https://www.gnu.org/licenses/gpl-3.0.en.html
 //
-// Time-stamp: <Wednesday 2024-07-31 05:59:04 +1000 Graham Williams>
+// Time-stamp: <Tuesday 2024-08-06 13:40:21 +1000 Graham Williams>
 //
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the GNU General Public License as published by the Free Software
@@ -92,7 +92,7 @@ class _SummaryDisplayState extends ConsumerState<SummaryDisplay> {
           This is the most basic R command for summarising the dataset.
           
           For **numeric data** the minimum, and maximum values are listed.
-          Between these we can see listed the first and thrid quartiles as well
+          Between these we can see listed the first and third quartiles as well
           as the median (the second quartile) and the mean.
           
           For **categoric data** a frequency table is provided, showing the
@@ -187,19 +187,36 @@ class _SummaryDisplayState extends ConsumerState<SummaryDisplay> {
     ////////////////////////////////////////////////////////////////////////
 
     content = 'Kurtosis:\n';
-    content += rExtract(stdout, 'kurtosis(ds[numc], na.rm=TRUE)');
+    content += rExtract(stdout, 'timeDate::kurtosis(ds[numc], na.rm=TRUE)');
     content += '\nSkewness:\n';
-    content += rExtract(stdout, 'skewness(ds[numc], na.rm=TRUE)');
+    content += rExtract(stdout, 'timeDate::skewness(ds[numc], na.rm=TRUE)');
 
-    // Add some spacing to the output.
+    // Regular expression to match lines like '[1] "xxxx"'
+
+    final pattern = RegExp(r'^\[1\] "(.*)"$', multiLine: true);
+
+    // Replace matched lines with 'calculated using method="xxxx"'
+
+    content = content.replaceAllMapped(pattern, (match) {
+      String method = match.group(1) ?? '';
+      return 'Calculated using method="$method"';
+    });
+
+    // Iterate over the lines and modify them.
 
     // Split the string into lines.
 
     lines = content.split('\n');
 
+    // Filter out lines that match 'attr(,"method")'
+
+    lines = lines.where((line) => line.trim() != 'attr(,"method")').toList();
+
     // Iterate over the lines and modify them.
 
     for (int i = 0; i < lines.length; i++) {
+      // Add some spacing to the output.
+
       if (!RegExp(r'^\s*[\d-]').hasMatch(lines[i])) {
         lines[i] = '\n${lines[i]}';
       }
@@ -212,9 +229,16 @@ class _SummaryDisplayState extends ConsumerState<SummaryDisplay> {
     if (lines.length > 4) {
       pages.add(
         TextPage(
-          title: '# Kurtosis and Skewness\n\n'
-              'Generated using [fBasics::kurtosis(ds) and fBasics::skewness(ds)]'
-              '(https://www.rdocumentation.org/packages/fBasics).\n\n',
+          title: '''
+
+          # Kurtosis and Skewness
+
+          Generated using
+          [timeDate::kurtosis(ds)](https://www.rdocumentation.org/packages/timeDate/topics/kurtosis)
+          and
+          [timeDate::skewness(ds)](https://www.rdocumentation.org/packages/timeDate/topics/skewness).
+
+          ''',
           content: '\n$content',
         ),
       );
@@ -224,7 +248,7 @@ class _SummaryDisplayState extends ConsumerState<SummaryDisplay> {
     // FBASICS STATS
     ////////////////////////////////////////////////////////////////////////
 
-    content = rExtract(stdout, 'lapply(ds[numc], basicStats)');
+    content = rExtract(stdout, 'lapply(ds[numc], fBasics::basicStats)');
 
     // Replace $ at the beginning of any lines.
 
@@ -237,6 +261,42 @@ class _SummaryDisplayState extends ConsumerState<SummaryDisplay> {
           title: '# Detailed Variable Statistics\n\n'
               'Generated using [fBasics::basicStats(ds)]'
               '(https://www.rdocumentation.org/packages/fBasics/topics/BasicStatistics).\n\n',
+          content: '\n$content',
+        ),
+      );
+    }
+
+    ////////////////////////////////////////////////////////////////////////
+    // CROSSTAB<
+    ////////////////////////////////////////////////////////////////////////
+
+    content = rExtract(stdout, 'descr::CrossTable(ds');
+
+    // Remove any line beginning with +.
+
+    lines = content.split('\n');
+
+    // Iterate over the lines and modify them.
+
+    // Remove lines starting with '+'
+
+    lines = lines.where((line) => !line.trimLeft().startsWith('+')).toList();
+
+    // Join the lines back into a single string.
+
+    content = lines.join('\n');
+
+    if (content.isNotEmpty) {
+      pages.add(
+        TextPage(
+          title: '''
+
+          # Cross Tabulation
+
+          Generated using
+          [descr::CrossTable(ds)](https://www.rdocumentation.org/packages/descr/topics/CrossTable)
+
+              ''',
           content: '\n$content',
         ),
       );