+++ lib/features/tests/config.dart
@@ -5,7 +5,7 @@
 /// License: GNU General Public License, Version 3 (the "License")
 /// https://www.gnu.org/licenses/gpl-3.0.en.html
 //
-// Time-stamp: <Tuesday 2024-07-30 19:49:58 +1000 Graham Williams>
+// Time-stamp: <Sunday 2024-08-04 07:44:01 +1000 Graham Williams>
 //
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the GNU General Public License as published by the Free Software
@@ -50,7 +50,7 @@ class TestsConfigState extends ConsumerState<TestsConfig> {
   @override
   Widget build(BuildContext context) {
     // update the rolesProvider to get the latest inputs
-    updateRolesProvider(ref);
+    updateVariablesProvider(ref);
     // Retrieve the current selected variable and use that as the initial value
     // for the dropdown menu. If there is no current value then we choose the
     // first input variable.
@@ -86,7 +86,7 @@ class TestsConfigState extends ConsumerState<TestsConfig> {
                 rSource(context, ref, 'test');
                 // showUnderConstruction(context);
               },
-              child: const Text('Display'),
+              child: const Text('Perform Statistical Tests'),
             ),
             const SizedBox(
               width: 10,