+++ lib/providers/pty.dart
@@ -1,6 +1,6 @@
 /// A provider of the pseudo terminal running R.
 ///
-/// Time-stamp: <Tuesday 2024-06-04 09:45:48 +1000 Graham Williams>
+/// Time-stamp: <Saturday 2024-08-03 20:54:20 +1000 Graham Williams>
 ///
 /// Copyright (C) 2023, Togaware Pty Ltd.
 ///
@@ -61,7 +61,7 @@ final ptyProvider = StateProvider<Pty>((ref) {
 
   terminal.onOutput = (data) {
     // This gets called when a user types into the R console. So typing the
-    // command `ls()` in the R conosle the command is echoed in the R
+    // command `ls()` in the R console the command is echoed in the R
     // console. This is not capturing the output from the console.
 
     pty.write(const Utf8Encoder().convert(data));
@@ -74,11 +74,15 @@ final ptyProvider = StateProvider<Pty>((ref) {
   return pty;
 });
 
-/// We are only interested in running R on whichever desktop.
+/// We are interested in running R on whichever desktop.
 ///
 /// Linux and MacOS desktops initiate R simply through the R command. Windows
 /// does an R.exe.
 
 String get shell {
-  return Platform.isWindows ? 'R.exe' : 'R';
+  return Platform.isWindows
+      ? 'R.exe'
+      : Platform.isMacOS
+          ? '/usr/local/bin/R'
+          : '/usr/bin/R';
 }